First steps to LoD heightmap
In particular, we've set up SimpleHeightmap to be ready for multiple LoD tiles
(it just needs to generate them on update_lod) and swapped indices down to u16
since large heightmaps will be cut up into tiles.

+++ src/model/mem.rs
@@ -25,7 +25,7 @@ pub struct Geometry {
 	/// The object's vertexes.
 	pub vertices: Vec<Vertex>,
 	/// The object's geometry, specified by indexes into the vertex vector.
-	pub indices: Vec<u32>,
+	pub indices: Vec<u16>,
 }
 
 /// In-memory material and texture specification.
